refactoed/cleaned up code.

+++ JSports/site/src/Services/BulletinService.php
@@ -27,15 +27,16 @@ use Joomla\CMS\Factory;
 use Joomla\CMS\Component\ComponentHelper;
 use Joomla\CMS\Uri\Uri;
 use Joomla\Filesystem\Folder;
+// use Joomla\CMS\Filesystem\Path;
 use FP4P\Component\JSports\Site\Services\LogService;
 use FP4P\Component\JSports\Site\Objects\Application as Myapp;
 
 class BulletinService
 {
    
-    public static function getBulletinsTable()
+    public static function getBulletinsTable() : BulletinsTable
     {
-        $db = Factory::getDbo();
+        $db = Factory::getContainer()->get(DatabaseInterface::class);
         return new BulletinsTable($db);
     }
     
@@ -45,44 +46,60 @@ class BulletinService
      * @param number $id
      * @return \FP4P\Component\JSports\Administrator\Table\BulletinsTable|NULL
      */
-    public static function getItem($id = 0) {
+    public static function getItem(int $id = 0) : ?BulletinsTable {
         
-        $db = Factory::getDbo();
+        $db = Factory::getContainer()->get(DatabaseInterface::class);
         $bulletins = new BulletinsTable($db);
                 
         $row = $bulletins->load($id);
         
         if ($row) {
             return $bulletins;
-        }
-               
+        }               
         return null;
     }
 
-    public static function delete($id = 0) {
+    /**
+     * This function will delete an individual bulletin record in the database.
+     * NOTE:  deleting any underlying attachment is handled elsewhere.
+     * 
+     * @param int $id
+     * @return bool
+     */
+    public static function delete(int $id = 0) : bool {
         
         $logger = Myapp::getLogger();
+        if ($id === 0) {
+            $logger->error('Bulletin Record ID ' . $id . ' is required ');
+            return false;
+        }
         
         $svc = new BulletinService();
         $item = $svc->getItem($id);
+        if ($item === null) {
+            $logger->error("Bulletin record not found (id=$id)");
+            return false;
+        }
         
-        $db = Factory::getDbo();
-        
+        $db = Factory::getContainer()->get(DatabaseInterface::class);
         $query = $db->getQuery(true);
         
         $conditions = array(
-            $db->quoteName('id') . '=' .$db->quote($id));
+            $db->quoteName('id') . ' = :id');
         
         $query->delete($db->quoteName('#__jsports_bulletins'));
         $query->where($conditions);
-        
+        $query->bind(':id', $id, ParameterType::INTEGER);
         $db->setQuery($query);
         
-        $rc = $db->execute();
-        $logger->info('Deleting bullet item - ' . $item->title . ' ID: ' . $id);
-        
-        return $rc;
-        
+        try {
+            $db->execute();
+            $logger->info('Deleting bulletin item - ' . $item->title . ' ID: ' . $id);
+            return true;
+        } catch (\Throwable $e) {
+            $logger->error('Bulletin delete failed - Record ID ' . $id . ' - ' . $e->getMessage());
+            return false;
+        }
     }
     
     /**
@@ -93,9 +110,9 @@ class BulletinService
      * @param string $key
      * @return string
      */
-    public static function getBulletinFilePath($key) {
+    public static function getBulletinFilePath(int $key) : string {
         $params = ComponentHelper::getParams('com_jsports');
-        $attachmentdir = $params->get('attachmentdir');
+        $attachmentdir = trim((string) $params->get('attachmentdir',''));
         
         $value = rtrim($attachmentdir); // optional: remove trailing whitespace
         
@@ -114,13 +131,22 @@ class BulletinService
         return $filepath;
     }
     
-    public static function getBulletinAttachmentURL($key, $filename) {
+    /**
+     * This function will return the URL string that can be used to retrieve the bulletins
+     * attachment via a browser.
+     * 
+     * @param string $key Same as the bulletinid
+     * @param string $filename
+     * @return string
+     */
+    public static function getBulletinAttachmentURL(int $key, string $filename) : string {
         $params = ComponentHelper::getParams('com_jsports');
-        $attachmentdir = $params->get('attachmentdir');
+        $attachmentdir = trim((string) $params->get('attachmentdir',''));
         
-        $filepath =  Uri::root() . $attachmentdir . '/Bulletin-' . $key .'/' . $filename;
+//         $filepath =  Uri::root() . $attachmentdir . '/Bulletin-' . $key .'/' . $filename;
+//         return $filepath;
+        return Uri::root() . $attachmentdir . '/Bulletin-' . $key . '/' . rawurlencode($filename);
         
-        return $filepath;
     }
     
     
@@ -129,7 +155,7 @@ class BulletinService
      * @param string $key
      * @return boolean
      */
-    public static function deleteAttachmentFolder($key) {
+    public static function deleteAttachmentFolder(int $key) {
         
         $filepath = BulletinService::getBulletinFilePath($key);
         
@@ -151,8 +177,15 @@ class BulletinService
         
     }
 
-    
-    public static function clearAttachmentFilename($id) {
+    /**
+     * This function will blank out the attachment filenmae for a specific bulletin.
+     * @param int $id
+     * @return bool
+     */
+    public static function clearAttachmentFilename(int $id) : bool {
+        if ($id === 0) {
+            return false;
+        }
         return BulletinService::updateAttachmentFilename($id, "");    
     }
     
@@ -163,8 +196,8 @@ class BulletinService
      * @param string $name
      * @return boolean
      */
-    public static function updateAttachmentFilename($id, $name) {
-//         $db = Factory::getDbo();
+    public static function updateAttachmentFilename(int $id, string $name) : bool {
+
         $db = Factory::getContainer()->get(DatabaseInterface::class);
         $query = $db->getQuery(true);
         
@@ -182,17 +215,18 @@ class BulletinService
             ->set($fields)
             ->where($conditions);
             
-            $query->bind(':bulletinid', $id, ParameterType::INTEGER)->bind(':filename_value', $name, ParameterType::STRING);
+            $query->bind(':bulletinid', $id, ParameterType::INTEGER)
+                ->bind(':filename_value', $name, ParameterType::STRING);
             
             $db->setQuery($query);
-            
             $db->execute();
             return true;
+            
         } catch (\Exception $e) {
             LogService::error($e->getMessage());
             return false;
         }
-        return true;
+//         return true;
     }
 }
